renamed folders and removed unnecessary package
