+++ server/package.json
@@ -14,8 +14,7 @@
     "body-parser": "^1.20.0",
     "dotenv": "^16.0.2",
     "express": "^4.18.1",
-    "node-cron": "^3.0.2",
-    "shelljs": "^0.8.5"
+    "node-cron": "^3.0.2"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
